// src/model/kernel/pessoa/errors/PessoaFuncionarioErrorModel.tsx
export type PessoaFuncionarioErrorModel = {
    pesNome?: string;
    pesMae?: string;
    pesEmail?: string;
    pesNascimento?: string;
    pesStatus?: string;
    pesEstadocivil?: string;
    pesNacionalidade?: string;
    pesTelefone?: string;
    pesCpf?: string;
    pesDeficiencia?: string;
    tpPessoa?: string;
    areaAtuacao?: string;
}