// src/model/kernel/pessoa/PessoaFisicaModel.tsx
export type PessoaFisicaModel = {
    pesId?: number;
    pesNome?: string;
    pesMae?: string;
    pesEmail?: string;
    pesNascimento?: string; //Date
    pesStatus?: string;
    pesEstadocivil?: string;
    pesNacionalidade?: string;
    pesTelefone?: string;
    pesCpf?: string;
    pesDeficiencia?: number;
    tpPessoa?: number;
}